+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Version 1.1
 
 * Task, command, and settings for cleaning up Python cache files and build artifacts (#4)
+* New setting for Pip's `--upgrade-strategy` option (#3)
 
 
 ## Version 1.0